+++ pytermor/__main__.py
@@ -2,34 +2,52 @@
 #  pytermor [ANSI formatted terminal output toolset]
 #  (c) 2022-2023. A. Shavykin <0.delameter@gmail.com>
 # -----------------------------------------------------------------------------
-from os.path import dirname, join
+from __future__ import annotations
+
+import typing as t
 from subprocess import PIPE, run, CalledProcessError, DEVNULL
+from os import stat
+from os.path import dirname, join
+from datetime import datetime
 
-try:
-    gitdir = dirname(__file__) + "a"
-    cp = run(
-        ["git", "-C", gitdir, "describe", "--tags"],
-        stdout=PIPE,
-        stderr=DEVNULL,
-        check=True,
-    )
-    version = cp.stdout.strip().decode()
-    cp = run(
-        ["git", "-C", gitdir, "show", "-s", "--format=%cd", "--date=format:%b-%y"],
-        stdout=PIPE,
-        stderr=DEVNULL,
-        check=True,
-    )
-    dt = cp.stdout.strip().decode()
+from ._version import __version__
 
-except CalledProcessError as e:
-    from datetime import datetime
-    from os import stat
-    from ._version import __version__
 
-    version = __version__
+class Main:
+    def __init__(self):
+        _iter_fns: t.List[t.Callable] = [
+            self._get_version_from_git,
+            self._get_version_from_fs,
+        ]
+        while len(_iter_fns) > 0:
+            fn = _iter_fns.pop(0)
+            if (result := fn()) is not None:
+                print("v" + ":".join(result))
+                return
+        print("Failed to determine the version")
 
-    versionfile = join(dirname(__file__), "_version.py")
-    dt = datetime.fromtimestamp(stat(versionfile).st_mtime).strftime("%b-%y")
+    def _call_git(self, *args) -> str:
+        gitdir = dirname(__file__)
+        cp = run(["git", "-C", gitdir, *args], stdout=PIPE, stderr=DEVNULL, check=True)
+        return cp.stdout.strip().decode()
 
-print(f"v{version}:{dt}")
+    def _get_version_from_git(self) -> t.Tuple[str, str]|None:
+        try:
+            v = self._call_git("describe", "--tags")
+            dt = self._call_git("showw", "-s", "--format=%cd", "--date=format:%b-%y")
+            return v, dt
+        except (CalledProcessError, UnicodeDecodeError):
+            return None
+
+    def _get_version_from_fs(self) -> t.Tuple[str, str]|None:
+        try:
+            vfile = join(dirname(__file__), "_version.py")
+            mtime = stat(vfile).st_mtime
+            dt = datetime.fromtimestamp(mtime).strftime("%b-%y")
+            return __version__, dt
+        except FileNotFoundError:
+            return None
+
+
+if __name__ == "__main__":
+    Main()